improved

+++ khomichenko.alisa/P4/incPer.cpp
@@ -9,22 +9,22 @@ void khomichenko::doInc(size_t start, size_t inc, size_t rows, int * matrix)
     return;
   }
   int stop = start;
-  for (int i = start; i != rows; ++i)
+  for (size_t i = start; i != rows; ++i)
   {
     matrix[i] += inc;
     stop = i;
   }
-  for(int i = stop+rows; i != stop+rows*(rows); i = i + rows)
+  for(size_t i = stop+rows; i != stop+rows*(rows); i = i + rows)
   {
     matrix[i] += inc;
     start = i;
   }
-  for(int i = start-1; i != start - (rows); --i)
+  for(size_t i = start-1; i != start - (rows); --i)
   {
     matrix[i] += inc;
     stop = i;
   }
-  for(int i = stop-rows; i > stop - rows * (rows - 1); i = i - rows)
+  for(size_t i = stop-rows; i > stop - rows * (rows - 1); i = i - rows)
   {
     matrix[i] += inc;
     stop = i;
@@ -35,7 +35,7 @@ void khomichenko::callFunc(int* matrix, const size_t cols, size_t rows)
 {
   size_t start = 0;
   size_t inc = 1;
-  int j = 0;
+  size_t j = 0;
   while (j + 1 != cols)
   {
     ++j;
